Split index into bundle & detect
