+++ packages/util-crypto/src/index.ts
@@ -1,31 +1,6 @@
 // Copyright 2017-2021 @polkadot/util-crypto authors & contributors
 // SPDX-License-Identifier: Apache-2.0
 
-import './cryptoInit';
 import './detectPackage';
 
-export { packageInfo } from './packageInfo';
-
-// all internal exports
-export * from './address';
-export * from './base32';
-export * from './base58';
-export * from './base64';
-export * from './blake2';
-export * from './crypto';
-export * from './ethereum';
-export * from './hd';
-export * from './hmac';
-export * from './json';
-export * from './keccak';
-export * from './key';
-export * from './mnemonic';
-export * from './nacl';
-export * from './pbkdf2';
-export * from './random';
-export * from './schnorrkel';
-export * from './scrypt';
-export * from './secp256k1';
-export * from './sha512';
-export * from './signature';
-export * from './xxhash';
+export * from './bundle';
