+++ packages/util/src/index.ts
@@ -3,25 +3,4 @@
 
 import './detectPackage';
 
-/**
- * @summary Utility methods for this package are split into groups
- */
-export { packageInfo } from './packageInfo';
-
-export * from './array';
-export * from './assert';
-export * from './bn';
-export * from './buffer';
-export * from './compact';
-export * from './extractTime';
-export * from './format';
-export * from './hex';
-export * from './is';
-export * from './logger';
-export * from './memoize';
-export * from './number';
-export * from './promisify';
-export * from './string';
-export * from './stringify';
-export * from './u8a';
-export * from './versionDetect';
+export * from './bundle';
